requirements.txt
Django>=4.2,<5
djangorestframework>=3.15
psycopg2-binary>=2.9
celery>=5.4
redis>=5.0
pandas>=2.2
openpyxl>=3.1
python-dateutil>=2.9
gunicorn>=21.2
python-dotenv>=1.0
